+++ habot/io/messages.py
@@ -1,79 +1,19 @@
 """
-Communications with non-API external entities.
-
-Currently this means interacting via private or party messages in Habitica.
+Handling for communications via Habitica messages.
 """
 
-from collections import OrderedDict
 from datetime import datetime
-from io import StringIO
-from lxml import etree
 import requests.exceptions
-import yaml
 
-from habitica_helper.habiticatool import PartyTool
-from habitica_helper.task import Task
 from habitica_helper.utils import get_dict_from_api, timestamp_to_datetime
 from habitica_helper import habrequest
 
 from conf.tasks import PM_SENT, GROUP_MSG_SENT
-from habot.db import DBOperator
+from habot.io.db import DBOperator
 from habot.exceptions import CommunicationFailedException
 from habot.habitica_operations import HabiticaOperator
 import habot.logger
 from habot.message import PrivateMessage, ChatMessage, SystemMessage
-
-
-class DBTool():
-    """
-    High-level tools for using the database.
-    """
-    # pylint: disable=too-few-public-methods
-
-    def __init__(self):
-        """
-        Initialize the class
-        """
-        self._logger = habot.logger.get_logger()
-        self._db = DBOperator()
-
-    def get_user_id(self, habitica_loginname):
-        """
-        Return the user ID of a party member corresponding to given login name.
-        """
-        members = self._db.query_table(
-            "members",
-            condition="loginname='{}'".format(habitica_loginname),
-            columns="id",
-            )
-        if not members:
-            raise ValueError("User with login name {} not found"
-                             "".format(habitica_loginname))
-        return members[0]["id"]
-
-    def get_party_user_ids(self):
-        """
-        Return a list of user IDs for all party members.
-        """
-        members = self._db.query_table(
-            "members",
-            columns="id",
-            )
-        return [data_dict["id"] for data_dict in members]
-
-    def get_loginname(self, uid):
-        """
-        Return the login name of the party member with the given UID.
-        """
-        members = self._db.query_table(
-            "members",
-            condition="id='{}'".format(uid),
-            columns="loginname",
-            )
-        if not members:
-            raise ValueError("User with user ID {} not found"
-                             "".format(uid))
-        return members[0]["loginname"]
 
 
 class HabiticaMessager():
@@ -469,300 +409,6 @@ class HabiticaMessager():
         return False
 
 
-class DBSyncer():
-    """
-    Fetch data from Habitica API and write it to the database.
-    """
-
-    def __init__(self, header):
-        """
-        :header: Habitica API call header for a party member
-        """
-        self._header = header
-        self._db = DBOperator()
-        self._logger = habot.logger.get_logger()
-
-    def update_partymember_data(self):
-        """
-        Fetch current party member data from Habitica and update the database.
-
-        If the database contains members that are not currently in the party,
-        they are removed from the database.
-        """
-        self._logger.debug("Going to update partymember data in the DB.")
-        partytool = PartyTool(self._header)
-        partymembers = partytool.party_members()
-
-        self.add_new_members(partymembers)
-        self._logger.debug("Added new members")
-        self.remove_old_members(partymembers)
-        self._logger.debug("Removed outdated members")
-
-    def remove_old_members(self, partymembers):
-        """
-        Remove everyone who is not a current party member from "members" table.
-
-        :partymembers: A complete list of current party members.
-        """
-        member_ids_in_party = [member.id for member in partymembers]
-        members_in_db = self._db.query_table("members", "id")
-        for member in members_in_db:
-            if member["id"] not in member_ids_in_party:
-                self._db.delete_row("members", "id", member["id"])
-
-    def add_new_members(self, partymembers):
-        """
-        Update the database to contain data for all given party members.
-
-        If someone is missing entirely, they are added, or if someone's
-        information has changed (e.g. displayname), the corresponding row is
-        updated.
-        """
-        for member in partymembers:
-            db_data = {
-                "id": member.id,
-                "displayname": member.displayname,
-                "loginname": member.login_name,
-                "birthday": member.habitica_birthday,
-                }
-            user_row = self._db.query_table(
-                "members", condition="id='{}'".format(member.id))
-            if len(user_row) == 0:
-                self._db.insert_data("members", db_data)
-            elif user_row != db_data:
-                self._db.update_row("members", member.id, db_data)
-
-
-class YAMLFileIO():
-    """
-    Read and write YAML files in a way that benefits the bot.
-    """
-
-    @classmethod
-    def read_tasks(cls, filename):
-        """
-        Create tasks representing the ones in the file.
-
-        The input file must be a YAML file, containing a list of dicts, each of
-        them representing one task. Each task must have keys "text" and
-        "tasktype", and may also contain any of the following: "notes", "date",
-        "difficulty", "uppable" and "downable".
-
-        :returns: A list of tasks
-        """
-        tasks = []
-        with open(filename) as taskfile:
-            file_contents = yaml.load(taskfile, Loader=yaml.BaseLoader)
-            for taskdict in file_contents:
-                # TODO error handling
-                tasks.append(Task(taskdict))
-        return tasks
-
-    @classmethod
-    def read_question_list(cls, filename, unused_only=False):
-        """
-        Parse a YAML file containing weekly questions.
-
-        The expected syntax for the file is that it contains a list
-        "questions", each item of which is a dict with keys "question",
-        "description" and "used". For example
-
-        questions:
-          - question: What is your favourite fruit?
-            description: Do you like bananas or apples more? Or maybe kiwis?
-            used: True
-          - question: What is your favourite animal?
-            description: The only real answer here is labrador though =3
-            used: False
-
-        is a valid question list file.
-
-        :returns: An OrderedDict of Tasks. Only the text, tasktype and notes
-                  are set for the task, everything else has to be added later.
-                  The value for each task is a boolean that denotes if the task
-                  was marked as being used already.
-        """
-        with open(filename) as questionfile:
-            file_contents = yaml.load(questionfile, Loader=yaml.BaseLoader)
-            try:
-                questions = file_contents["questions"]
-            except KeyError as key_error:
-                raise \
-                    MalformedQuestionFileException(
-                        "The question file doesn't seem to contain a question "
-                        "list", filename) \
-                    from key_error
-            question_tasks = OrderedDict()
-            for question in questions:
-                try:
-                    if unused_only and question["used"].lower() == "true":
-                        continue
-
-                    task_data = {
-                        "text": question["question"],
-                        "tasktype": "todo",
-                        "notes": question["description"],
-                        }
-                    question_tasks[Task(task_data)] = (
-                        question["used"].lower() == "true")
-                except KeyError as key_error:
-                    raise \
-                        MalformedQuestionFileException(
-                            "The following question in the question list is "
-                            "malformed:\n{}".format(question),
-                            filename) \
-                        from key_error
-
-            return question_tasks
-
-    @classmethod
-    def write_question_list(cls, questions, filename):
-        """
-        Save all questions as YAML into the given file.
-
-        The given questions must be a dict, keys of which are Tasks and values
-        booleans determining whether that question has already been used.
-
-        questions:
-          - question: What is your favourite fruit?
-            description: Do you like bananas or apples more? Or maybe kiwis?
-            used: True
-          - question: What is your favourite animal?
-            description: The only real answer here is labrador though =3
-            used: False
-
-        :questions: A dict of Habitica tasks and booleans telling whether they
-                    have already been used in some previous challenge.
-        :filename: The output file.
-        """
-        question_data = []
-        for question in questions:
-            question_data.append({
-                "question": question.text,
-                "description": question.notes,
-                "used": questions[question]})
-        with open(filename, "w") as dest:
-            yaml.dump({"questions": question_data}, dest,
-                      default_flow_style=False)
-
-
-class WikiReader():
-    """
-    Tool for fetching content of a page from Habitica wiki.
-    """
-
-    def __init__(self, url):
-        """
-        Initialize the object
-
-        :url: String containing url to the page, e.g.
-              "https://habitica.fandom.com/wiki/The_Keep:Mental_Health_Warriors_Unite".
-        """
-        self.url = url
-        self._parser = etree.HTMLParser()
-        self._page = None
-
-    class Decorators():
-        """
-        Decorators used by WikiReader
-        """
-        # pylint: disable=too-few-public-methods
-
-        @classmethod
-        def needs_page(cls, method):
-            """
-            Decorator for functions that need the page to be fetched
-            """
-            def _decorated(*args, **kwargs):
-                # pylint: disable=protected-access
-                if not args[0]._page:
-                    args[0]._read_page()
-                return method(*args, **kwargs)
-            return _decorated
-
-    def _read_page(self):
-        """
-        Fetch the page from the wiki.
-
-        :raise:
-            :HTTPError: if the page cannot be fetched
-            :WikiParsingError: if the page content cannot be found
-        """
-        response = requests.get(self.url)
-        response.raise_for_status()
-        full_page_tree = etree.parse(StringIO(str(response.content)),
-                                     self._parser)
-        content = full_page_tree.getroot().cssselect(".WikiaPage")
-        if len(content) != 1:
-            raise WikiParsingError("More than one `WikiaPage` element "
-                                   "encountered")
-        self._page = content[0]
-
-    @property
-    @Decorators.needs_page
-    def page(self):
-        """
-        Return the HTML contents of the page as lxml ElementTree.
-        :returns: lxml Element corresponding to the root node of the wiki page
-                  content.
-        """
-        return self._page
-
-    @Decorators.needs_page
-    def find_elements_with_matching_subelement(self, element_selector,
-                                               child_text):
-        """
-        Return a list of elements of given type that have a matching children.
-
-        The given `child_text` must only be present in one of the children
-        elements: it is not necessary for it to be the full content of the
-        element.
-
-        :element_selector: CSS selector for finding the elements. E.g. `div` or
-                           `#someid` or `ul.navigation`.
-        :child_text: A string that must be found in the content of at least one
-                     of the child elements.
-        :returns: A list of lxml ElementTrees, each startig from an element
-                  that matched the search criteria.
-        """
-        css_matching_elements = self._page.cssselect(element_selector)
-        full_match_elements = []
-        for element in css_matching_elements:
-            for child in element.iterchildren():
-                if child.text and child_text in child.text:
-                    full_match_elements.append(element)
-        return full_match_elements
-
-
-class MalformedQuestionFileException(Exception):
-    """
-    Exception raised when the question list cannot be parsed.
-
-    Has the following attributes:
-    :problem: A short description of the problem
-    :filename: The problematic file
-    :message: A custom extra message
-    """
-
-    _INFO = ('The expected syntax for the file is that it contains a list '
-             '"questions", each item of which is a dict with keys "question", '
-             '"description" and "used". For example:\n\n'
-             'questions:\n'
-             '  - question: What is your favourite fruit?\n'
-             '    description: Do you like bananas or apples more? Or maybe '
-             'kiwis?\n'
-             '    used: True\n'
-             '  - question: What is your favourite animal?\n'
-             '    description: The only real answer here is labrador ofc =3\n'
-             '    used: False\n\n'
-             'is a valid question list file.')
-
-    def __init__(self, problem, filename):
-        message = ("Problem with question file \"{}\":\n\n{}\n\n{}"
-                   "".format(filename, problem, self._INFO))
-        super().__init__(message)
-
-
 class UnsplittableMessage(Exception):
     """
     Exception raised when splitting a too long message is not possible.
@@ -772,10 +418,4 @@ class UnsplittableMessage(Exception):
 class SpamDetected(Exception):
     """
     Exception for situations where the spot is being used for spamming.
-    """
-
-
-class WikiParsingError(Exception):
-    """
-    Exception for unexpected content from Habitica Wiki.
     """
