+++ tests/io/db_test.py
@@ -1,17 +1,16 @@
 """
-Test `habot.io` module.
+Test `habot.io.db` module.
 """
 
 import datetime
-from unittest import mock
 
 import pytest
 
 from habitica_helper.habiticatool import PartyTool
 from habitica_helper.member import Member
 
-from habot.db import DBOperator
-from habot.io import HabiticaMessager, DBSyncer, DBTool, WikiReader
+from habot.io.db import DBOperator, DBSyncer, DBTool
+from habot.io.messages import HabiticaMessager
 
 
 @pytest.fixture()
@@ -20,26 +19,6 @@ def test_messager(header_fx):
     Create a HabiticaMessager for testing purposes.
     """
     return HabiticaMessager(header_fx)
-
-
-# pylint doesn't understand fixtures
-# pylint: disable=redefined-outer-name
-
-@mock.patch("habitica_helper.habrequest.post")
-@mock.patch("habot.habitica_operations.HabiticaOperator.tick_task")
-def test_group_message(mock_tick, mock_post, test_messager, header_fx):
-    """
-    Test group message sending.
-
-    Ensure that a correct API call is made to send a group message with the
-    message in its payload, and that a task is ticked afterwards.
-    """
-    test_messager.send_group_message("group-id", "some message")
-    mock_post.assert_called_with(
-            "https://habitica.com/api/v3/groups/group-id/chat",
-            headers=header_fx,
-            data={"message": "some message"})
-    mock_tick.assert_called()
 
 
 @pytest.fixture
@@ -51,6 +30,9 @@ def db_operator_fx(db_connection_fx):
     # pylint: disable=unused-argument
     yield DBOperator()
 
+
+# Pylint doesn't understand fixtures
+# pylint: disable=redefined-outer-name
 
 PARTY_CHAT_MSG_1 = {
         "flagCount": 0,
@@ -108,7 +90,7 @@ def patch_get_dict_response(monkeypatch):
         # pylint: disable=unused-argument
         def _return_messages(*args, **kwargs):
             return messages
-        monkeypatch.setattr("habot.io.get_dict_from_api",
+        monkeypatch.setattr("habot.io.messages.get_dict_from_api",
                             _return_messages)
     return _patch
 
@@ -462,55 +444,3 @@ def test_get_non_existent_login_name(db_tool_fx):
         db_tool_fx.get_loginname("nonexistent-member-uid")
     assert ("User with user ID nonexistent-member-uid not found"
             in str(err.value))
-
-
-@pytest.fixture
-def patch_wiki_page(requests_mock):
-    """
-    Make get request to wiki return a constant article from a file.
-
-    The page has a all the versatile content Mental Health Warriors party page
-    has.
-    """
-    with open("tests/data/party-wikipage.html") as htmlfile:
-        page = htmlfile.read()
-    requests_mock.get("https://habitica.fandom.com/wiki/test_article",
-                      text=page)
-
-
-@pytest.mark.usefixtures("patch_wiki_page")
-def test_get_wiki_page():
-    """
-    Test using WikiReader to get a page.
-
-    Ensure that it has real content elements from start and end of the page.
-    """
-    reader = WikiReader("https://habitica.fandom.com/wiki/test_article")
-    assert reader.page
-    assert ("The Keep:Mental Health Warriors Unite" in
-            [h.text for h in reader.page.xpath("//h1")])
-    assert (reader.page.xpath("//span")[-1].text ==
-            "Take your favorite fandoms with you and never miss a beat.")
-
-
-@pytest.mark.usefixtures("patch_wiki_page")
-def test_find_elements_with_matching_subelement():
-    """
-    Test that a list of elements with the correct content is returned.
-
-    The method is currently only tested using a query that returns a single
-    element, but it is ensured that the element is of the right type and has
-    the expected content.
-    """
-    reader = WikiReader("https://habitica.fandom.com/wiki/test_article")
-    quest_list = reader.find_elements_with_matching_subelement("ul",
-                                                               "(CURRENT)")
-    assert len(quest_list) == 1
-    assert quest_list[0].tag == "ul"
-
-    li_elements = quest_list[0].getchildren()
-    assert len(li_elements) == 8
-    assert li_elements[0].text == "(CURRENT) Robot (collection)"
-
-    for element in li_elements:
-        assert element.tag == "li"
